requirements.txt
flask
flask-httpauth
pyserial
Pillow
